Only create initial versions after cloning a document.
