+++ opengever/document/subscribers.py
@@ -1,10 +1,12 @@
 from five import grok
+from OFS.interfaces import IObjectClonedEvent
 from opengever.document import _
 from opengever.document.behaviors import IBaseDocument
 from opengever.document.document import IDocumentSchema
 from opengever.ogds.base.utils import ogds_service
 from plone.i18n.normalizer.interfaces import IIDNormalizer
 from zope.component import getUtility
+from zope.component.hooks import getSite
 from zope.globalrequest import getRequest
 from zope.i18n import translate
 from zope.lifecycleevent.interfaces import IObjectAddedEvent
@@ -12,6 +14,23 @@ from zope.lifecycleevent.interfaces import IObjectCopiedEvent
 from zope.lifecycleevent.interfaces import IObjectCreatedEvent
 from zope.lifecycleevent.interfaces import IObjectModifiedEvent
 import os.path
+
+
+@grok.subscribe(IDocumentSchema, IObjectClonedEvent)
+def create_initial_version(obj, event):
+    """When a object was copied, create an initial version.
+    """
+    portal = getSite()
+    pr = portal.portal_repository
+    history = pr.getHistory(obj)
+
+    if history is not None and not len(history) > 0:
+        comment = translate(_(u'label_initial_version_copied',
+                            default="Initial version (document copied)"),
+                            context=getRequest())
+        # Create an initial version
+        pr._recursiveSave(obj, {}, pr._prepareSysMetadata(comment),
+            autoapply=pr.autoapply)
 
 
 @grok.subscribe(IDocumentSchema, IObjectAddedEvent)
